Safari-611.1.10/LayoutTests/fast/shadow-dom/pointerlockelement-in-slot-expected.txt

Test pointerLockElement is the node in the same tree as the context object.
To manually test, click on "Click here" below.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.pointerLockElement is null
PASS shadowRoot.pointerLockElement is null
PASS document.pointerLockElement is target
PASS shadowRoot.pointerLockElement is null
PASS anotherShadowRoot.pointerLockElement is null
PASS document.pointerLockElement is null
PASS shadowRoot.pointerLockElement is null
PASS anotherShadowRoot.pointerLockElement is null
PASS successfullyParsed is true

TEST COMPLETE
Click here
